+++ runtime/common/src/lib.rs
@@ -175,6 +175,13 @@ pub mod consts {
 pub mod config {
 	use super::*;
 
+	pub mod common {
+		use super::*;
+		parameter_types! {
+		pub const TreasuryPalletId: PalletId = PalletId(*b"py/trsry");
+		}
+	}
+
 pub mod frame_system{
 	use super::*;
 
@@ -228,4 +235,17 @@ pub mod pallet_timestamp{
 	}
 }
 
+pub mod pallet_treasury {
+	use super::*;
+
+	parameter_types! {
+		pub const ProposalBond: Permill = Permill::from_percent(5);
+		pub const ProposalBondMinimum: Balance = 1 * currency::DOLLARS;
+		pub const ProposalBondMaximum: Option<Balance> = None;
+		pub const SpendPeriod: BlockNumber = 1 * consts::DAYS;
+		pub const Burn: Permill = Permill::from_percent(0);
+		pub const MaxApprovals: u32 = 100;
+	}
+
+}
 }
